+++ lib/api.js
@@ -1,4 +1,4 @@
-import { FrequentMetrics, InfrequentMetrics } from "../lib/constants";
+import { FrequentMetrics, InfrequentMetrics } from "../lib/metrics";
 
 function constructQuery(start, period, events) {
   const query = {
